requirements.txt
requests 
cryptography 
python-dotenv